refactor: moved constants
