+++ cognite_toolkit/_cdf_tk/templates/__init__.py
@@ -1,22 +1,8 @@
-from ._constants import (
-    BUILD_ENVIRONMENT_FILE,
-    COGNITE_MODULES,
-    COGNITE_MODULES_PATH,
-    CUSTOM_MODULES,
-    ROOT_MODULES,
-    ROOT_PATH,
-)
 from ._utils import flatten_dict, iterate_modules, module_from_path, resource_folder_from_path
 
 __all__ = [
     "iterate_modules",
     "module_from_path",
     "resource_folder_from_path",
-    "COGNITE_MODULES",
-    "CUSTOM_MODULES",
-    "ROOT_PATH",
-    "ROOT_MODULES",
-    "COGNITE_MODULES_PATH",
-    "BUILD_ENVIRONMENT_FILE",
     "flatten_dict",
 ]
