--- cognite_toolkit/_cdf_tk/templates/_constants.py
@@ -1,25 +1,0 @@
-from pathlib import Path
-
-# This is the default config located locally in each module.
-DEFAULT_CONFIG_FILE = "default.config.yaml"
-# The environment file:
-BUILD_ENVIRONMENT_FILE = "_build_environment.yaml"
-
-COGNITE_MODULES = "cognite_modules"
-CUSTOM_MODULES = "custom_modules"
-ALT_CUSTOM_MODULES = "modules"
-
-ROOT_MODULES = [COGNITE_MODULES, CUSTOM_MODULES, ALT_CUSTOM_MODULES]
-MODULE_PATH_SEP = "/"
-
-# Add any other files below that should be included in a build
-EXCL_FILES = ["README.md", DEFAULT_CONFIG_FILE]
-# Which suffixes to exclude when we create indexed files (i.e., they are bundled with their main config file)
-EXCL_INDEX_SUFFIX = frozenset([".sql", ".csv", ".parquet"])
-# Files to search for variables.
-SEARCH_VARIABLES_SUFFIX = frozenset([".yaml", "yml", ".sql", ".csv"])
-# Which suffixes to process for template variable replacement
-PROC_TMPL_VARS_SUFFIX = frozenset([".yaml", ".yml", ".sql", ".csv", ".parquet", ".json", ".txt", ".md", ".html", ".py"])
-
-ROOT_PATH = Path(__file__).parent.parent.parent
-COGNITE_MODULES_PATH = ROOT_PATH / COGNITE_MODULES
