+++ tests/tests_unit/test_cli/test_command_sequences.py
@@ -15,7 +15,8 @@ import typer
 from pytest import MonkeyPatch
 
 from cognite_toolkit._cdf import build, clean, deploy, main_init
-from cognite_toolkit._cdf_tk.templates import COGNITE_MODULES, iterate_modules
+from cognite_toolkit._cdf_tk.constants import COGNITE_MODULES
+from cognite_toolkit._cdf_tk.templates import iterate_modules
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig
 from tests.constants import REPO_ROOT
 from tests.tests_unit.approval_client import ApprovalCogniteClient
